requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.3
distlib==0.3.1
Django==2.2.16
django-ckeditor==6.1.0
django-ckeditor-5==0.0.14
django-js-asset==1.2.2
django-tenant-schemas==1.10.0
django-tenants==3.3.2
django-widget-tweaks==1.4.8
dropbox==11.19.0
filelock==3.0.12
idna==3.2
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
ordered-set==4.0.2
Pillow==8.3.2
ply==3.11
psycopg2==2.9.1
psycopg2-binary==2.8.6
pycparser==2.20
pylint==2.6.0
pytz==2020.1
PyYAML==5.3.1
requests==2.26.0
six==1.15.0
sqlparse==0.3.1
stone==3.2.1
toml==0.10.1
urllib3==1.26.6
virtualenv==20.0.31
whitenoise==5.2.0
wrapt==1.12.1
